DocAI/requirements_complete.txt
# Core Web Framework
flask>=2.3.3
flask-cors>=4.0.0
werkzeug>=2.3.7
gunicorn>=20.1.0

# AI/ML and LLM Libraries
groq>=0.4.0
pocketgroq>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-huggingface>=0.0.3
langchain-groq>=0.1.0
openai>=1.12.0

# Document Processing Libraries
docx2txt>=0.8
python-docx>=0.8.11
PyPDF2>=3.0.0
pypdf>=3.0.0

# Vector Database and Embeddings
faiss-cpu>=1.7.4

# Text Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# HTTP Requests
requests>=2.31.0

# System Process Management
psutil>=5.9.0

# Image Processing
Pillow>=10.0.0

# Development and Testing
pydantic>=2.0.0

# Machine Learning and Transformers
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2

# Security
cryptography>=41.0.0

# Note: LibreOffice UNO bindings are linked from system installation
# System LibreOffice must be installed separately